+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add `external-structure` theme that colors any geometry by structure properties
 - Support float and half-float data type for direct-volume rendering and GPU isosurface extraction
 - Minor documentation updates
+- Add support for position-location to `volume-value` color theme
+- Add support for color themes to `slice` representation
+- Improve/fix palette support in volume color themes
 - Support removal of independent selection controls in the viewport
 - Support custom selection controls
 - Fix `Plane3D.projectPoint`